Move platform/host to platform-os
