+++ platform-os/src/udp-receive.cpp
@@ -4,7 +4,7 @@
 // This file is subject to the terms and conditions defined in the
 // file 'LICENSE', which is part of this source code package.
 //------------------------------------------------------------------------------
-#include "beetree/platform/host/udp-receive.hpp"
+#include "beetree/platform/udp-receive.hpp"
 
 #include <cstdio>
 #include <cstring>
